requirements.txt
spoon-ai-sdk>=0.3.0
spoon-toolkits>=0.2.0
python-dotenv>=1.0.0
